V11.10.04: rewrite Fertilizer.java, WateringCan.java, and move StoreAble.java from entity to use cases.
