+++ app/src/main/java/com/farmgame/entity/Item/Fertilizer.java
@@ -1,19 +1,30 @@
 package com.farmgame.entity.Item;
 
+import static com.farmgame.constants.Constants.Type_Fertilizer;
 import com.farmgame.entity.LandEntity;
+import com.farmgame.usecase.StoreAble;
 
-public class Fertilizer implements Item{
+
+
+public class Fertilizer implements Item, StoreAble {
     /**
      * Fertilizer is a class that implements Item interface, it has a fixed price, which is 10. Its
-     * purpose is to fertilize a land if necessary.
+     * purpose is to fertilize a land if required. It also implements the store-able interface.
      */
     private final int price;
-    private final String itemname;
+    private final String item_name;
 
+
+
+    /**
+     * Constructor for Fertilizer.
+     */
     public Fertilizer(){
         this.price = 10;
-        this.itemname = "Fertilizer";
+        this.item_name = Type_Fertilizer;
     }
+
+
 
     /**
      * A getter method for the price of fertilizer.
@@ -25,10 +36,17 @@ public class Fertilizer implements Item{
         return this.price;
     }
 
+
+
+    /**
+     * A getter method for the name of fertilizer.
+     *
+     * @return the name of fertilizer.
+     */
     @Override
-    public String getName(){
-        return this.itemname;
-    }
+    public String getName(){ return this.item_name; }
+
+
 
     /**
      * Fertilize the land with fertilizer.
